InMemoryRangeIndex Also Add:  - txn_graph and txn_graph_all virtual tables for debugging transaction dependency graphs  - txn_cache virtual table to debug accord cache contents Also Fix:  - NPE in AccordDebugKeyspace.TxnOpsTable  - AccordCoordinatorMetrics latency metrics should use Timer not Histogram  - OnDiskIndex.readLast Also Improve:  - Catchup should run after progress log is started  - Cancelling AccordTask should early-terminate range scanning  - Improve RangeTxnScanner description  - Add ballot to txn tables  - Add NodeId to logging output
patch by Benedict; reviewed by Alex Petrov for CASSANDRA-21306

+++ src/java/org/apache/cassandra/metrics/DecayingEstimatedHistogramReservoir.java
@@ -579,18 +579,18 @@ public class DecayingEstimatedHistogramReservoir implements CassandraReservoir
                 throw new IllegalStateException("Unable to compute when histogram overflowed");
 
             long elements = 0;
-            long sum = 0;
+            double sum = 0;
             for (int i = 0; i < lastBucket; i++)
             {
                 long bCount = decayingBuckets[i];
                 elements += bCount;
-                sum += bCount * bucketOffsets[i];
+                sum += bCount * (double)bucketOffsets[i];
             }
 
             if (elements == 0)
                 return 0d;
 
-            return (double) sum / elements;
+            return sum / elements;
         }
 
         /**
